news/requirements.txt
streamlit
psycopg2-binary
ollama
